refactor: station 패키지명 수정
